delete unneccessary files

+++ user_view.py
@@ -15,16 +15,18 @@ def home():
     current_datetime = datetime.now()
 
     categories_nav = EventCategory.query.all()
-    upcoming_event = Event.query.filter(Event.event_start >= current_datetime).all()
-    event_images = Event.query.filter(Event.event_img != None).all()
+    upcoming_event = Event.query.filter(Event.event_start >= current_datetime).filter_by(publish_status="Published").all()
+    new_event = Event.query.filter(Event.event_start >= current_datetime).filter_by(publish_status="Published").order_by(Event.last_updated.desc()).all()
+    event_images = Event.query.filter(Event.event_img != None, Event.event_start >= current_datetime).filter_by(publish_status="Published").all()
 
-    return render_template('Homepage.html', categories_nav=categories_nav, events=upcoming_event, event_images=event_images, user=current_user)
+    return render_template('home.html', categories_nav=categories_nav, upcoming=upcoming_event, new=new_event,
+                            event_images=event_images, user=current_user)
 
 @user_view.route('/events', defaults={'category': None})
 @user_view.route('/events/<category>')
 @login_required
 def event_category(category):
-    # Retrieve search query and venue filter from URL query parameters
+    # Retrieve value from URL query parameters
     search_query = request.args.get('search', '').strip()
     venue_filter = request.args.get('venue_type', '').strip()
     sort = request.args.get('sort', '').strip()
@@ -69,7 +71,7 @@ def event_category(category):
     events_pagination = event_query.paginate(page=page, per_page=9)
     events = events_pagination.items
 
-    return render_template('event-category.html', category=category_obj, event=events, categories_nav=categories_nav,
+    return render_template('event_category.html', category=category_obj, event=events, categories_nav=categories_nav,
                             current_datetime=current_datetime, search_query=search_query, venue_filter=venue_filter, 
                             sort=sort, pagination=events_pagination, include_expired=include_expired, user=current_user)
 
@@ -107,15 +109,17 @@ def event_details(event_id):
     # Get current datetime
     current_datetime = datetime.now()
 
-    return render_template('EventDetails.html', event=event, user_has_ticket=user_has_ticket,
+    return render_template('event_details.html', event=event, user_has_ticket=user_has_ticket,
                            tickets_info=tickets_info, current_datetime=current_datetime, categories_nav=categories_nav, user=current_user)
 
+#after choose the ticket type, user will be bring to order summary with order details
 @user_view.route('/order_summary', methods=['GET'])
 @login_required
 def order_summary():
     # Navigation bar
     categories_nav = EventCategory.query.all()
 
+    # Retrieve value from URL query parameters
     event_name = request.args.get('event_name')
     ticket_id = request.args.get('ticket_id')
     ticket_type = request.args.get('ticket_type')
@@ -124,6 +128,34 @@ def order_summary():
     return render_template('order_summary.html', categories_nav=categories_nav, event_name=event_name, ticket_id=ticket_id, 
                            ticket_type=ticket_type, ticket_price=ticket_price, user=current_user)
 
+# for free tickets
+@user_view.route('/confirm_order', methods=['POST'])
+@login_required
+def confirm_order():
+    try:
+        ticket_id = request.form.get('ticket_id')
+
+        # Create a new order entry
+        new_order = UserOrder(
+            user_id=current_user.user_id,
+            ticket_id=ticket_id,
+            order_date=datetime.now(),
+            order_quantity=1,
+            total_amount=0,  # As the price is 0.00
+            invoice="Free"
+        )
+
+        db.session.add(new_order)
+        db.session.commit()
+
+        return render_template('payment_success.html')  # Adjust the redirect URL as needed
+
+    except Exception as e:
+        flash(f'Error confirming order: {str(e)}', 'error')
+        return redirect(url_for('user_view.payment_cancel'))
+
+
+#details will be listed again on checkout page for double confirmation
 @user_view.route('/checkout', methods=['POST'])
 @login_required
 def checkout():
@@ -145,7 +177,7 @@ def checkout():
                 'product_data': {
                     'name': f"{event_name} - {ticket_type}",
                 },
-                'unit_amount': int(float(ticket_price) * 100),
+                'unit_amount': int(float(ticket_price) * 100), # Stripe stores amounts in cents
             },
             'quantity': 1,
         }],
@@ -162,12 +194,14 @@ def save_order_to_db(session):
         ticket_id = session['client_reference_id']
         user_id = current_user.user_id
         order_quantity = 1  
+        total_amount = session['amount_total'] / 100  
 
         new_order = UserOrder(
             user_id=user_id,
             ticket_id=ticket_id,
             order_date=datetime.now(),
             order_quantity=order_quantity,
+            total_amount=total_amount,
             invoice=session['id']
         )
 
@@ -178,7 +212,7 @@ def save_order_to_db(session):
         flash(f'Error saving order to database: {str(e)}', 'error')
         return redirect(url_for('user_view.home'))  # Adjust the redirect URL as needed
 
-
+# If payment is successful, show success and redirect to homepage
 @user_view.route('/payment_success')
 @login_required
 def payment_success():
@@ -198,6 +232,7 @@ def payment_success():
 
     return redirect(url_for('home'))
 
+#If payment is cancelled, show cancel and redirect to homepage
 @user_view.route('/payment_cancel')
 @login_required
 def payment_cancel():
@@ -240,32 +275,44 @@ def organisers():
 def manage_membership(admin_id, action):
     user_id = current_user.user_id
 
-    if action == 'request':
-        existing_membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id, mstatus_id=1).first()
-        if not existing_membership:
-            new_membership = Membership(
-                user_id=user_id,
-                admin_id=admin_id,
-                mstatus_id=1  # 1 represents Pending
-            )
-            db.session.add(new_membership)
-            db.session.commit()
-            flash('Membership request submitted!', 'success')
+    if request.method == 'POST':
+        action = request.form.get('action')
 
-    elif action == 'unjoin':
-        membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id, mstatus_id=2).first()
-        if membership:
-            db.session.delete(membership)
-            db.session.commit()
-            flash('Successfully unjoined membership.', 'success')
-    elif action == 'cancel':
-        membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id, mstatus_id=1).first()
-        if membership:
-            db.session.delete(membership)
-            db.session.commit()
-            flash('Membership request cancelled.', 'success')
+        if action == 'request':
+            existing_membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id).first()
 
-    return redirect(url_for('user_view.organisers'))
+            if not existing_membership:
+                # If no membership exists, create a new one with status 'Pending'
+                new_membership = Membership(
+                    user_id=user_id,
+                    admin_id=admin_id,
+                    mstatus_id=1  # 1 represents 'Pending'
+                )
+                db.session.add(new_membership)
+                db.session.commit()
+                flash('Membership request submitted!', 'success')
+            else:
+                # If membership exists, update its status to 'Pending' if it's not already 'Pending'
+                if existing_membership.mstatus_id != 1:  # Check if status is not 'Pending'
+                    existing_membership.mstatus_id = 1
+                    db.session.commit()
+                    flash('Membership request submitted!', 'success')
+
+        elif action == 'unjoin':
+            membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id, mstatus_id=2).first()
+            if membership:
+                db.session.delete(membership)
+                db.session.commit()
+                flash('Successfully unjoined membership.', 'success')
+        
+        elif action == 'cancel':
+            membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id, mstatus_id=1).first()
+            if membership:
+                db.session.delete(membership)
+                db.session.commit()
+                flash('Membership request cancelled.', 'success')
+
+    return redirect(url_for('user_view.organisers', admin_id=admin_id))
 
 @user_view.route('/organisers/details/<admin_id>', methods=['GET'])
 @login_required
@@ -284,10 +331,14 @@ def organiser_details(admin_id):
     # Get current datetime
     current_datetime = datetime.now()
 
-    # Fetch events
-    all_events = Event.query.filter_by(admin_id=admin_id, publish_status="Published").order_by(Event.event_start.desc()).all()
+    # Fetch events with pagination
+    page = request.args.get('page', 1, type=int)
+    per_page = 9  # Define the number of items per page
+    events_pagination = Event.query.filter_by(admin_id=admin_id, publish_status="Published").order_by(Event.event_start.desc()).paginate(page=page, per_page=per_page)
+    events = events_pagination.items
 
-    return render_template('organiser_details.html', organiser=admin, categories_nav=categories_nav, events=all_events, current_datetime=current_datetime, user=current_user)
+    return render_template('organiser_details.html', organiser=admin, categories_nav=categories_nav, events=events, current_datetime=current_datetime, 
+                           pagination=events_pagination, user=current_user)
 
 @user_view.route('/organisers/details/<admin_id>/<action>', methods=['POST'])
 @login_required
@@ -298,16 +349,26 @@ def details_manage_membership(admin_id, action):
         action = request.form.get('action')
 
         if action == 'request':
-            existing_membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id, mstatus_id=1).first()
+            existing_membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id).first()
+
             if not existing_membership:
+                # If no membership exists, create a new one with status 'Pending'
                 new_membership = Membership(
                     user_id=user_id,
                     admin_id=admin_id,
-                    mstatus_id=1  # 1 represents Pending
+                    mstatus_id=1  # 1 represents 'Pending'
                 )
                 db.session.add(new_membership)
                 db.session.commit()
                 flash('Membership request submitted!', 'success')
+            else:
+                # If membership exists, update its status to 'Pending' if it's not already 'Pending'
+                if existing_membership.mstatus_id != 1:  # Check if status is not 'Pending'
+                    existing_membership.mstatus_id = 1
+                    db.session.commit()
+                    flash('Membership status updated to Pending.', 'success')
+                else:
+                    flash('Membership request already exists.', 'info')
 
         elif action == 'unjoin':
             membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id, mstatus_id=2).first()
@@ -315,6 +376,7 @@ def details_manage_membership(admin_id, action):
                 db.session.delete(membership)
                 db.session.commit()
                 flash('Successfully unjoined membership.', 'success')
+        
         elif action == 'cancel':
             membership = Membership.query.filter_by(user_id=user_id, admin_id=admin_id, mstatus_id=1).first()
             if membership:
@@ -331,7 +393,7 @@ def user_profile(user_id):
     user_id = current_user.user_id
     
     user = User.query.get_or_404(user_id)
-    form = ResetPasswordForm()
+    form = ResetPasswordForm() # Get form
     categories_nav = EventCategory.query.all() # Event category in navigation bar
     return render_template('user_profile.html', user=user, form=form, categories_nav=categories_nav)
 
@@ -342,26 +404,22 @@ def reset_password():
     
     form = ResetPasswordForm()
     user_id = request.args.get('user_id')  # Extract user_id from request arguments
-    print(f"Received request to reset password for user_id: {user_id}")
     
     if form.validate_on_submit():
-        print("Form validated successfully.")
         user = User.query.get_or_404(user_id)
-        
+
+        # Check hashed old password
         if bcrypt.check_password_hash(user.user_pwd, form.old_pwd.data):
-            print("Old password matched.")
             new_hashed_password = bcrypt.generate_password_hash(form.new_pwd.data).decode('utf-8')
             user.user_pwd = new_hashed_password
             db.session.commit()
-            flash('Your password has been reset!', 'success')
+            flash('Password reset successfully!', 'success')
         else:
-            print("Old password did not match.")
             flash('Old password is incorrect', 'danger')
     else:
-        print("Form validation failed.")
         for field, errors in form.errors.items():
             for error in errors:
-                flash(f"Error in {getattr(form, field).label.text}: {error}", 'danger')
+                flash(f"Wrong value in '{getattr(form, field).label.text}': {error}", 'danger')
 
     return redirect(url_for('user_view.user_profile', user_id=user_id))
